maj type to Object pour les attributs, ajout des liens entre les classes

+++ backend/src/main/java/com/cpweb/backend/repositories/IdentifiantRepository.java
@@ -1,8 +1,7 @@
 package com.cpweb.backend.repositories;
 
-import com.cpweb.backend.models.Categorie;
 import com.cpweb.backend.models.Identifiant;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface IdentifiantRepository extends JpaRepository<Identifiant, Integer> {
+public interface IdentifiantRepository extends JpaRepository<Identifiant, Long> {
 }
