+++ backend/src/main/java/com/cpweb/backend/repositories/IngredientRepository.java
@@ -1,8 +1,7 @@
 package com.cpweb.backend.repositories;
 
-import com.cpweb.backend.models.Categorie;
 import com.cpweb.backend.models.Ingredient;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface IngredientRepository extends JpaRepository<Ingredient, Integer> {
+public interface IngredientRepository extends JpaRepository<Ingredient, Long> {
 }
