+++ backend/src/main/java/com/cpweb/backend/repositories/InstructionRepository.java
@@ -1,8 +1,7 @@
 package com.cpweb.backend.repositories;
 
-import com.cpweb.backend.models.Categorie;
 import com.cpweb.backend.models.Instruction;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface InstructionRepository extends JpaRepository<Instruction, Integer> {
+public interface InstructionRepository extends JpaRepository<Instruction, Long> {
 }
