+++ backend/src/main/java/com/cpweb/backend/repositories/TagRecetteRepository.java
@@ -1,8 +1,7 @@
 package com.cpweb.backend.repositories;
 
-import com.cpweb.backend.models.Categorie;
 import com.cpweb.backend.models.TagRecette;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TagRecetteRepository extends JpaRepository<TagRecette, Integer> {
+public interface TagRecetteRepository extends JpaRepository<TagRecette, Long> {
 }
