+++ backend/src/main/java/com/cpweb/backend/repositories/UtilisateurRepository.java
@@ -1,8 +1,7 @@
 package com.cpweb.backend.repositories;
 
-import com.cpweb.backend.models.Categorie;
 import com.cpweb.backend.models.Utilisateur;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UtilisateurRepository extends JpaRepository<Utilisateur, Integer> {
+public interface UtilisateurRepository extends JpaRepository<Utilisateur, Long> {
 }
